Makefile and norme

+++ eleonard_asm/srcs/fill_by_zeroes.c
@@ -6,7 +6,7 @@
 /*   By: eleonard <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/29 15:02:54 by eleonard          #+#    #+#             */
-/*   Updated: 2020/01/29 15:02:55 by eleonard         ###   ########.fr       */
+/*   Updated: 2020/01/31 17:57:51 by eleonard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 void	fill_by_zeroes(char *str, int len)
 {
-	size_t	i;
-	size_t	code_len;
+	int	i;
+	int	code_len;
 
 	i = 0;
-	code_len = ft_strlen(str);
+	code_len = (int)ft_strlen(str);
 	while (i < len)
 	{
 		str[code_len + i] = '0';
